+++ tests/testthat/test_spwb.R
@@ -208,3 +208,20 @@ test_that("spwb_day gives same result with inner and direct calls with general c
   expect_equal(medfate::copy_model_output(ic, x3i, "spwb"), s_dir)
 
 })
+
+test_that("spwb_day does not produce warnings with truncated root distirbution",{
+  exampleforest$shrubData$Z50 <- 100
+  exampleforest$shrubData$Z95 <- 200
+  control_granier$truncateRootDistribution<- TRUE
+  x1 <- spwbInput(exampleforest, examplesoil, SpParamsMED, control_granier)
+  expect_s3_class(spwb_day(x1, date, meteovec,
+                           latitude = 41.82592, elevation = 100, slope=0, aspect=0), "spwb_day")
+  control_sperry$truncateRootDistribution<- TRUE
+  x2 <- spwbInput(exampleforest, examplesoil, SpParamsMED, control_sperry)
+  expect_s3_class(spwb_day(x2, date, meteovec,
+                           latitude = 41.82592, elevation = 100, slope=0, aspect=0), "spwb_day")
+  control_sureau$truncateRootDistribution<- TRUE
+  x3 <- spwbInput(exampleforest, examplesoil, SpParamsMED, control_sureau)
+  expect_s3_class(spwb_day(x3, date, meteovec,
+                           latitude = 41.82592, elevation = 100, slope=0, aspect=0), "spwb_day")
+})